LPS-86601 Set home page function moved to LayoutUtils
